requirements.txt
PyPDF2==3.0.1
python-docx==1.1.2
nltk==3.8.1
requests==2.31.0